Move PackageOptions methods to Package
